This commit introduces a new, non-breaking feature: error correlation. Destination-specific errors can now be received on the corresponding destination observables as terminal errors.
The user can provide a callback called `correlateErrors`. This callback is called against every error frame. If the callback returns the ID of a currently subscribed destination, the error frame is emitted as an error on the corresponding observables, terminating them. In this manner, better use of the native observable API is made for clients working against brokers that may issue destination-specific errors. Also, we ensure a disconnect-reconnect loop is avoided by preventing the automatic re-subscription of a destination that produced an error.

Some brokers may wish to produce destination-specific errors. For example, a broker might check the users token on SUBSCRIBE to a particular destination for correct destination-specific privileges, sending an ERROR if the required privileges are absent. In this case, the STOMP spec advises the server to mention the offending destination ID somewhere in the frame, without mandating exactly how (https://stomp.github.io/stomp-specification-1.2.html#ERROR).

Without this error correlation feature, the error is only emitted from a general purpose `stompErrors$` observable and the error callbacks of the related observables are unused. This means that the related observable remains open, causing a re-SUBSCRIBE to be issued to its destination when the socket automatically re-connects. This can result in a tight loop of error-drop-reconnect, effectively bricking the entire socket for all channels.

* `RxStomp` constructor modified to make it possible to inject the stomp client to use. This makes unit testing scenarios requiring slightly more complex broker behavior much easier (it could be argued that the rest of the tests are really _integration_ tests).
* `karmaTypescriptConfig` property `instrumentation` set to false. Without this I couldn't get accurate source mapping while debugging unit tests in Chrome.
* Changing the protocol for `karma-safari-launcher` from `git` to `https` to fix `npm ci` error.